feat(cavatica): SKFP-956 add sorter when starting cavatica analyse
